backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0
dashscope>=1.14.0
numpy>=1.24.0
scikit-learn>=1.3.0
websockets>=12.0
python-dotenv>=1.0.0
